Change tenant collection to repository.
