+++ server/routerlicious/packages/routerlicious-base/package.json
@@ -128,6 +128,13 @@
 		}
 	},
 	"typeValidation": {
-		"broken": {}
+		"broken": {
+			"InterfaceDeclaration_ITenantDocument": {
+				"forwardCompat": false
+			},
+			"InterfaceDeclaration_ITenantDocument": {
+				"backCompat": false
+			}
+		}
 	}
 }
