+++ server/routerlicious/packages/routerlicious-base/src/riddler/app.ts
@@ -3,7 +3,12 @@
  * Licensed under the MIT License.
  */
 
-import { ISecretManager, ICache, ICollection } from "@fluidframework/server-services-core";
+import {
+	ISecretManager,
+	ICache,
+	ICollection,
+	ITenantDocument,
+} from "@fluidframework/server-services-core";
 import { BaseTelemetryProperties } from "@fluidframework/server-services-telemetry";
 import * as bodyParser from "body-parser";
 import express from "express";
@@ -15,7 +20,6 @@ import {
 } from "@fluidframework/server-services-utils";
 import { catch404, getTenantIdFromRequest, handleError } from "../utils";
 import * as api from "./api";
-import { ITenantDocument } from "./tenantManager";
 
 export function create(
 	tenantsCollection: ICollection<ITenantDocument>,
